requirements.txt
altair==5.2.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
branca==0.7.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
entrypoints==0.4
et-xmlfile==1.1.0
Faker==22.6.0
favicon==0.7.0
folium==0.15.1
fonttools==4.47.2
gitdb==4.0.11
GitPython==3.1.41
htbuilder==0.6.2
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.4
matplotlib==3.8.2
mdurl==0.1.2
more-itertools==10.2.0
numerize==0.12
numpy==1.26.3
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
pillow==10.2.0
plotly==5.18.0
protobuf==4.25.2
pyarrow==15.0.0
pydeck==0.8.1b0
Pygments==2.17.2
pymdown-extensions==10.7
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.4
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
six==1.16.0
smmap==5.0.1
soupsieve==2.5
st-annotated-text==4.0.1
st-pages==0.4.5
streamlit==1.30.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.3.6
streamlit-faker==0.0.3
streamlit-folium==0.18.0
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.2
streamlit-option-menu==0.3.12
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.2.0
validators==0.22.0
watchdog==3.0.0
xyzservices==2023.10.1
zipp==3.17.0
